Moved category examples to Category/; added Isomorphism
