+++ src/Category.hs
@@ -12,9 +12,8 @@ module Category where
 import qualified Prelude as P
 import Prelude hiding (id,(.))
 import GHC.Types (Constraint)
-import GHC.Generics (Generic,Generic1,(:*:),(:.:),Par1)
+import GHC.Generics (Generic,Generic1)
 import qualified Control.Arrow as A
-import Data.Constraint ((:-)(..),Dict(..),(\\),refl,trans)
 import Data.Distributive
 import Data.Functor.Rep
 
@@ -271,131 +270,3 @@ instance Representable r => ComonoidalR (->) r RepAnd where
 instance Representable r => CocartesianR (->) r RepAnd where
   ins = tabulate RepAnd
   jams (RepAnd _ a) = a
-
--------------------------------------------------------------------------------
--- | Constraint entailment
--------------------------------------------------------------------------------
-
-instance Category (:-) where
-  type Obj' (:-) = Unconstrained
-  id = refl
-  (.) = trans
-
--- Constraint (,) is not first class with GHC, so define a synonymous class.
-class    (p,q) => p && q
-instance (p,q) => p && q
-
-instance Monoidal (:-) (&&) where
-  p *** q = Sub (Dict \\ p \\ q)
-
-instance Cartesian (:-) (&&) where
-  exl = Sub Dict
-  exr = Sub Dict
-  dup = Sub Dict
-
--- The Monoidal and Cartesian operations are lifted from Data.Constraint
--- equivalents but those types involve (,) instead of (&&), which would lead to
--- kind errors in instance declarations.
-
--- Constraint entailment could be cocartesian but isn't in GHC (presumably
--- because GHC instance search doesn't backtrac). On the other hand, entaliment
--- can probably be closed, now that GHC supports implication constraints.
-
--------------------------------------------------------------------------------
--- | Opposite category
--------------------------------------------------------------------------------
-
-newtype Op k a b = Op { unOp :: b `k` a }
-
-instance Category k => Category (Op k) where
-  type Obj' (Op k) = Obj k
-  id = Op id
-  Op g . Op f = Op (f . g)
-
-instance Comonoidal k co => Monoidal (Op k) co where
-  Op f *** Op g = Op (f +++ g)
-
-instance Cocartesian k co => Cartesian (Op k) co where
-  exl = Op inl
-  exr = Op inr
-  dup = Op jam
-
-instance Monoidal k p => Comonoidal (Op k) p where
-  Op f +++ Op g = Op (f *** g)
-
-instance Cartesian k p => Cocartesian (Op k) p where
-  inl = Op exl
-  inr = Op exr
-  jam = Op dup
-
-instance ComonoidalR k r p => MonoidalR (Op k) r p where
-  cross (fmap unOp -> fs) = Op (plus fs)
-
-instance MonoidalR k r p => ComonoidalR (Op k) r p where
-  plus (fmap unOp -> fs) = Op (cross fs)
-
-instance CocartesianR k r p => CartesianR (Op k) r p where
-  exs  = Op <$> ins
-  dups = Op jams
-
-instance CartesianR k r p => CocartesianR (Op k) r p where
-  ins  = Op <$> exs
-  jams = Op dups
-
-instance BiproductR k r p => BiproductR (Op k) r p
-
--------------------------------------------------------------------------------
--- | Representable indices as objects
--------------------------------------------------------------------------------
-
--- Trie functors
-class (Representable (Trie i), Rep (Trie i) ~ i) => HasTrie i where
-  type Trie i :: * -> *
-
-instance HasTrie () where
-  type Trie () = Par1
-instance C2 HasTrie i j => HasTrie (i :+ j) where
-  type Trie (i :+ j) = Trie i :*: Trie j
-instance C2 HasTrie i j => HasTrie (i :* j) where
-  type Trie (i :* j) = Trie i :.: Trie j
-
-newtype Indexed k i j = Indexed { unIndexed :: Trie i `k` Trie j }
-
-class    Obj k (Trie i) => ObjI k i
-instance Obj k (Trie i) => ObjI k i
-
-instance Category k => Category (Indexed k) where
-  type Obj' (Indexed k) = ObjI k
-  id = Indexed id
-  Indexed g . Indexed f = Indexed (g . f)
-
-instance Monoidal k (:*:) => Monoidal (Indexed k) (:+) where
-  Indexed f *** Indexed g = Indexed (f *** g)
-
-instance Cartesian k (:*:) => Cartesian (Indexed k) (:+) where
-  exl = Indexed exl
-  exr = Indexed exr
-  dup = Indexed dup
-
-instance Comonoidal k (:*:) => Comonoidal (Indexed k) (:+) where
-  Indexed f +++ Indexed g = Indexed (f +++ g)
-
-instance Cocartesian k (:*:) => Cocartesian (Indexed k) (:+) where
-  inl = Indexed inl
-  inr = Indexed inr
-  jam = Indexed jam
-
-instance Biproduct k (:*:) => Biproduct (Indexed k) (:+)
-
--- TODO: generalize from (:*:) and (:+).
-
-#if 0
-
-type RepX r i = Rep r :* i
-
-instance MonoidalR k r (:.:) => MonoidalR (Indexed k) r RepX where
-  cross (fmap unIndexed -> fs) = Indexed (cross fs)
-
--- Hm. Needs unsaturated type synonym
-
-#endif
